requirements.txt
requests==2.21.0
Flask==3.0.3
oauthlib==3.0.1
pyOpenSSL==19.0.0
Flask-Login==0.6.3
mysql-connector-python==8.3.0
oauth~=1.0.1
music21~=7.1.0
Authlib~=0.15.5